fix(Vercel): separate authOptions into another file
